+++ tests/plugin_test/.phan/config.php
@@ -152,5 +152,6 @@ return [
         'SimplifyExpressionPlugin',
         'UnsafeCodePlugin',
         'DeprecateAliasPlugin',
+        'PHPDocInWrongCommentPlugin',
     ],
 ];